pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plumise-agent"
version = "0.1.0"
description = "Distributed LLM inference agent for the Plumise blockchain"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Plumise Team"}]

dependencies = [
    # ML / Inference
    "torch>=2.2",
    "transformers>=4.40",
    "accelerate>=0.28",
    "safetensors",
    "sentencepiece",
    "tokenizers",
    "numpy",

    # gRPC
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.25",

    # Blockchain
    "web3>=6.0.0",
    "eth-account>=0.10.0",

    # HTTP
    "aiohttp>=3.9.0",
    "fastapi>=0.110",
    "uvicorn>=0.28",

    # Config / CLI
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
]
gpu = [
    "bitsandbytes>=0.42",
]

[project.scripts]
plumise-agent = "plumise_agent.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]
